TransactionLog trial

+++ src/main/java/komodocrypto/model/database/Exchange.java
@@ -1,14 +1,16 @@
 package komodocrypto.model.database;
 
+import java.math.BigDecimal;
+
 public class Exchange {
 
     int exchange_id;
     String exchange_name;
-    double transfer_fee;
-    double buy_fee;
-    double sell_fee;
+    BigDecimal transfer_fee;
+    BigDecimal buy_fee;
+    BigDecimal sell_fee;
 
-    public Exchange(int exchange_id, String exchange_name, double transfer_fee, double buy_fee, double sell_fee) {
+    public Exchange(int exchange_id, String exchange_name, BigDecimal transfer_fee, BigDecimal buy_fee, BigDecimal sell_fee) {
         this.exchange_id = exchange_id;
         this.exchange_name = exchange_name;
         this.transfer_fee = transfer_fee;
@@ -35,27 +37,27 @@ public class Exchange {
         this.exchange_name = exchange_name;
     }
 
-    public double getTransfer_fee() {
+    public BigDecimal getTransfer_fee() {
         return transfer_fee;
     }
 
-    public void setTransfer_fee(double transfer_fee) {
+    public void setTransfer_fee(BigDecimal transfer_fee) {
         this.transfer_fee = transfer_fee;
     }
 
-    public double getBuy_fee() {
+    public BigDecimal getBuy_fee() {
         return buy_fee;
     }
 
-    public void setBuy_fee(double buy_fee) {
+    public void setBuy_fee(BigDecimal buy_fee) {
         this.buy_fee = buy_fee;
     }
 
-    public double getSell_fee() {
+    public BigDecimal getSell_fee() {
         return sell_fee;
     }
 
-    public void setSell_fee(double sell_fee) {
+    public void setSell_fee(BigDecimal sell_fee) {
         this.sell_fee = sell_fee;
     }
 }
